running insert code with spring
